# api/docs/swagger.yaml
basePath: /v2
definitions:
...
  errorpkg.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      moreInfo:
        type: string
      status:
        type: integer
      userMessage:
        type: string
    type: object
...
  router.FileCopyOptions:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
  router.FileDeleteOptions:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
...
  router.FileMoveOptions:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
  router.FilePatchNameOptions:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  service.File:
    properties:
      createTime:
        type: string
      id:
        type: string
      isShared:
        type: boolean
      name:
        type: string
      parentId:
        type: string
      permission:
        type: string
      snapshot:
        $ref: '#/definitions/service.Snapshot'
      type:
        type: string
      updateTime:
        type: string
      workspaceId:
        type: string
    type: object
  service.FileList:
    properties:
      data:
        items:
          $ref: '#/definitions/service.File'
        type: array
      page:
        type: integer
      query:
        $ref: '#/definitions/service.FileQuery'
      size:
        type: integer
      totalElements:
        type: integer
      totalPages:
        type: integer
    type: object
...
  service.GroupPermission:
    properties:
      group:
        $ref: '#/definitions/service.Group'
      id:
        type: string
      permission:
        type: string
    type: object
...
info:
  contact: {}
  title: Voltaserve API
  version: 2.0.0
paths:
  /files:
    delete:
      description: Delete
      operationId: files_delete
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.FileDeleteOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Delete
      tags:
      - Files
    get:
      description: Get by Path
      operationId: files_get_by_path
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.File'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Get by Path
      tags:
      - Files
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create
      operationId: files_create
      parameters:
      - description: Type
        in: query
        name: type
        required: true
        type: string
      - description: Workspace ID
        in: query
        name: workspace_id
        required: true
        type: string
      - description: Parent ID
        in: query
        name: parent_id
        type: string
      - description: Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Create
      tags:
      - Files
  /files/{id}:
    get:
      description: Get
      operationId: files_get
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.File'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Get
      tags:
      - Files
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Patch
      operationId: files_patch
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Patch
      tags:
      - Files
  /files/{id}/copy:
    post:
      description: Copy
      operationId: files_copy
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.FileCopyOptions'
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Copy
      tags:
      - Files
  /files/{id}/count:
    get:
      description: Count
      operationId: files_get_count
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Count
      tags:
      - Files
  /files/{id}/group_permissions:
    get:
      description: Get Group Permissions
      operationId: files_get_group_permissions
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.GroupPermission'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Get Group Permissions
      tags:
      - Files
  /files/{id}/list:
    get:
      description: List
      operationId: files_list
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: Page
        in: query
        name: page
        type: string
      - description: Size
        in: query
        name: size
        type: string
      - description: Sort By
        in: query
        name: sort_by
        type: string
      - description: Sort Order
        in: query
        name: sort_order
        type: string
      - description: Query
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.FileList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: List
      tags:
      - Files
  /files/{id}/move:
    post:
      description: Move
      operationId: files_move
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.FileMoveOptions'
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Move
      tags:
      - Files
  /files/{id}/name:
    patch:
      description: Patch Name
      operationId: files_patch_name
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.FilePatchNameOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.File'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Patch Name
      tags:
      - Files
  /files/{id}/original{ext}:
    get:
      description: Download Original
      operationId: files_download_original
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Access Token
        in: query
        name: access_token
        required: true
        type: string
      - description: Extension
        in: query
        name: ext
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Download Original
      tags:
      - Files
...
swagger: "2.0"